create booking

+++ structs/requests/create_booking.go
@@ -2,6 +2,6 @@ package requests
 
 type CreateBooking struct {
 	Name    string `json:"name" validate:"required"`
-	EventId int    `json:"event_id"  validate:"required"`
+	EventId int64  `json:"event_id"  validate:"required"`
 	Total   int    `json:"total" validate:"required"`
 }
